requirements.txt
django==1.10.5
django-extensions==1.7.6
djangorestframework==3.4.6
drf-nested-routers==0.11.1
psycopg2==2.6.2
requests==2.11.1
ipython==5.1.0
ipdb==0.10.1
werkzeug==0.11.15
